Move v2 heart beat class to heartbeat package.
