+++ naming/src/main/java/com/alibaba/nacos/naming/core/v2/client/manager/impl/IpPortBasedClientManager.java
@@ -23,7 +23,7 @@ import com.alibaba.nacos.naming.core.v2.client.Client;
 import com.alibaba.nacos.naming.core.v2.client.impl.IpPortBasedClient;
 import com.alibaba.nacos.naming.core.v2.client.manager.ClientManager;
 import com.alibaba.nacos.naming.core.v2.event.client.ClientEvent;
-import com.alibaba.nacos.naming.healthcheck.ClientBeatUpdateTask;
+import com.alibaba.nacos.naming.healthcheck.heartbeat.ClientBeatUpdateTask;
 import com.alibaba.nacos.naming.misc.GlobalExecutor;
 import com.alibaba.nacos.naming.misc.Loggers;
 import com.alibaba.nacos.naming.misc.NamingExecuteTaskDispatcher;
